# Apply micro-optimisation for Match API